feat: sse 연결 로직 커스텀 훅으로 분리 및 환경변수 적용
- useNotificationSSE 훅 추가
- Layout 컴포넌트 내부 로직 정리
- VITE_API_BASE_URL 적용
- TODO: SSE 토큰 만료 시 302 리디렉션 문제는 백엔드에서 처리 필요
